tests/wpt/web-platform-tests/web-bundle/subresource-loading/script-static-element-with-base.https.tentative.sub.html

<!DOCTYPE html>
<title>WebBundle subresource loading for static elements with a base element</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md"
/>
<base href="../resources/wbn/static-element/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <script type="webbundle">
    {
      "source": "../static-element.wbn",
      "resources": [
        "https://{{domains[]}}:{{ports[https][0]}}/web-bundle/resources/wbn/static-element/resources/script.js",
        "https://{{domains[]}}:{{ports[https][0]}}/web-bundle/resources/wbn/static-element/resources/style.css",
        "https://{{domains[]}}:{{ports[https][0]}}/web-bundle/resources/wbn/static-element/resources/style-imported-from-file.css",
        "https://{{domains[]}}:{{ports[https][0]}}/web-bundle/resources/wbn/static-element/resources/style-imported-from-tag.css"
      ],
      "scopes": [
        "https://{{domains[]}}:{{ports[https][0]}}/web-bundle/resources/wbn/static-element/scopes/"
      ]
    }
  </script>
  <style type="text/css">
    @import 'resources/style-imported-from-tag.css';
    @import 'scopes/style-imported-from-tag.css';
  </style>
  <link href="resources/style.css" rel=stylesheet>
  <link href="scopes/style.css" rel=stylesheet>
  <script src="resources/script.js"></script>
  <script src="scopes/script.js"></script>
  <script src="out-of-scope/script.js"></script>

  <script>
    setup(() => {
      assert_true(HTMLScriptElement.supports('webbundle'));
    });
  </script>
  <script src="/web-bundle/subresource-loading/resources/static-element-with-test.js"></script>
</body>
